Successfully refactored OpenGL
